Critical Readme updates

+++ controller/controller.h
@@ -3,7 +3,7 @@ unsigned long currentTime = 0;
 unsigned long previousTime = 0;
 unsigned long sensorPreviousTime = 0;
 uint8_t frameCounter = 0;
-uint8_t itterations = 0;
+uint32_t itterations = 0;
 
 bool all_ready = false;
 bool armed = false;
